Ticket #1363: Update pre-cond of treat_one_relative_path_element_2
The parameter loop_index of the op treat_one_relative_path_element_2
shall not be greater than the size of BrowsePathSource. This condition
is required by the called operation get_BrowsePathSource. In addition,
this condition shows that the relative path, or more generally the
computation of BrowsePathResult, is based on an element in
BrowsePathSource.