Ticket #1363: Simplify spec of treat_one_relative_path_element
In the specification of post-cond of treat_one_relative_path_element,
the details related to the relation between an element in
BrowsePathTarget and the elements in BrowsePathSource are expressed. The
proof of PO related to this post-cond is complicated. Indeed, these
details can be removed from the post-cond because they can be specified
by :
- post-cond of called operations
- pre-cond of called operations

Therefore, these details are removed from post-cond of
treat_one_path_element. Pre-cond of treat_one_path_element_1,
treat_one_path_element_2 and treat_browse_result_one_source are updated.