Ticket #1363: Add inv linking isBrowseResultAllocated and isBrowseResultComputed
The abstract variable isBrowseResultAllocated is defined in the
component browse_treatment_result_bs, which is imported by
browse_treatment_i. In the componenet browse_treatment, the abstract
variable isBrowseResultComputed is defined to give the status of browse
result computation. A successful browse result computation implies that
memory allocation for registering browse results is also successful.