Ticket #1363: Lift isBrowseResultAllocated to translate_browse_path
The operation compute_browse_result is promoted from translate_browse_path_result
to translate_browse_path. The variable isBrowseResultAllocated is
used in the precond of this operation. As there is no linking invariant
allowing to prove the PO related to precond deduction, precond related
to isBrowseResultAllocated shall be lifted from browse_treatment to
translate_browse_path_result for the op compute_browse_result.
Consequently :
    - the abstract variable shall be lifted to the component
    translate_browse_path.
    - the precond related to isBrowseResultAllocated shall be copied from
    translate_browse_path_result to translate_browse_path for compute_browse_result.

The postcond of compute_browse_result is also updated to take into
account of modifications of isBrowseResultAllocated and
isBrowseResultComputed.