Ticket #1363: Add inv linking browse context and its status
Browse context is set and cleared for each browse path source in the
operation compute_browse_result_from_source. When entering the
operation, isBrowseValueContextDefined is FALSE. A link between this
status and browse context shall be established in order to ensure that
browse context stays unchanged (which is undefined) by this operation.

The linking inv is added into browse_treatment_context and lifted up to
the browse_treatment component.