Ticket #1363: Add valuation of isBrowseResultAllocated at service_set_view_i
At the level of service_set_view, "isBrowseResultAllocated = FALSE"
always holds. This condition is needed by precond of operations of
translate_browse_path, which is called by service_set_view_i.
Therefore, the condition is added into invariant of service_set_view_i
and into precond of local operations of this component.

This ensures that the associated memory is always freed after processing
the request.