Ticket #1363: Update postcond of treat_one_translate_browse_path
The operation treat_one_translate_browse_path updates BrowsePath_Res_X
for the browse path which is the input parameter of the operation. The
domain of BrowsePath_Res_X stays unchanged (and is equal to BrowsePaths).
This condition is needed to ensure that, after initialization of
BrowsePath_Res_X, any browse path belonging to BrowsePaths also belongs
to the domain of BrowsePath_Res_X.