Ticket #1363: Update precond of set_ResponseBrowseNext_BrowseResult
The invariant related to typing of ResponseBrowseNext_ResultReferences
is preserved by the op set_ResponseBrowseNext_BrowseResult only if the
input parameter p_browseResultReferences is determined. Therefore,
precond related to this input parameter is enhanced. This precond is
satisfied when the op set_ResponseBrowseNext_BrowseResult is called,
because the browse results are allocated.