Ticket #1363: Correct clear_browse_result
We change the code of this operation so that it becomes re-entrant: it
now can be called several times in a row.  This will make it easier to
ensure that allocated memory is always freed.

Note: The specification of this basic operation does not have any
precondition, so this function must be re-entrant to fulfill its
specification.