Ticket #1455: B model: deallocate service request when type is not managed
Deallocate the service request when type is not managed into bless_msg* operations.
This is done at this level for the following reasons:
The B model is not able to use deallocate operation which relies on a defined message in B model.
Moreover the service event manager is not able to identify this case with the return status code neither to treat it at higher level