Ticket #1532: Client wrapper: avoid unexpected connection / disconnection event at user level
When calling SOPC_ClientHelper_Connect/SOPC_ClientHelper_Disconnect function the connections events are expected to be inhibited during the blocking call.
Since it was not checked that the received event source of the state machine change was inhibited, it was possible for the corresponding event not to be inhibited unexpectedly.