Ticket #1457: Browse service: ensure CP context is cleared only once
Some issues were observed of a double free of this context without possibility to reproduce it.
Moreover the attentive manual review has not shown any possibility that case occurs.
The choice is made to use the B model state hasContinuationPoint which should at least avoid clear operation when not necessary.