Ticket #1584: Make t_ConcreteValueRank abstract
The value of t_ConcreteValueRank is NAT1 \/ {-1} which can not be the
value of a CONCRETE_CONSTANT (these must be integer intervals). So it
has been changed to an abstract constant.

All places where t_ConcreteValueRank was used to define the
implementation type have been changed to use t_ValueRank instead, while
retaining t_ConcreteValueRank to provide additional conditions.