Ticket #1602: Remove s_session from session_core_2
s_session is abstract and we do not need this variable at this level of
abstraction.  It is equal to dom(a_state) so I can remove it from the
specification and replace it by the value in the file.

The glueing invariant is added to the implementation of session_core_1.