Ticket #1548: endpoints config: keep endpoints in low level structure to remain accessible
When using SOPC_CommonHelper_GetConfiguration and accessing the low level server config structure, the endpoints array was empty when using the high level API.
Keep a copy of the endpoints here by simply avoiding low level API to clear them in order they remain accessible as the rest of the server config through the structure.

Note: server config structure is meant ot be part of the low level API subject to change in the future.
The requested elements and ways to access to server config through high level API should be added to #1598 to defined a dedicated API matching the actual needs.