Ticket #1602: Strengthen precondition of check_server_session_user_auth_attempts
We need this precondition because in the implementation, we increase
a_server_user_auth_attempts(p_session) by 1.
Without it, we can have a value higher than the max int.
Moreover, I changed the definition of a_server_user_auth_attempts in the
specification to stay consistent with the invariant.