Ticket #1602: Fix loop invariant of find_session_to_close
The previous invariant was far from provable. Firstly, it was lacking
typing of local variables. Then, it was trying to state some properties
about the timestamps associated to sessions.

We could have modelled that formally, by also specifying what is a
session that is too recent to be considered. But the investment was too
high, with a payback too low.

So, it has been decided to keep the code with the timestamps, but to
simplify the loop invariant to just say either we found something
reasonable or not. This new version is sufficient to prove the
refinement.

The IF after the loop has been changed to test the session variable,
rather than its timestamp (the two are morally equivalent, but the loop
invariant does not say so). The new code is more readable in my opinion.