Ticket #1602: Add new operations in session_core_2
Previously, we were not resetting the count of user attempts when
closing a session. This could have been proved correct with the addition
of some clever invariants. But it would not be very robust, and it would
make debugging more complicated (you have to look at the state of the
session to find out whether the count of attempts is relevant or not).

Now we always reset the count of attempts to zero when closing a session.