Ticket #1602: Rework of server_activate_session_req_and_resp_sm
The previous specification was difficult to understand and it was
difficult to make the link between the implementation and the
specification.

This way of specification is longer but is more readable.

Moreover, I removed a test in the implementation because if the state of
the session is orphaned, the channel associated will never be a
s_channel_connected. So the test returned True everytime.