Ticket #1602: Rework of two operations in session_core
Firstly, I add a precondition in both operation to have the information if
the channel is a server or a client one.

Secondly, I removed some server variables from the client operation, and
some client variables from the server operation.

Finaly, I have replaced a LET by an ANY which makes proof easier.