Ticket #1602: Fix specification of client_create_session_req_sm
The specification was not correct, according to the implementation, we
have a channel and the creating state associated to the session, no
matter if valid is True or False.

Moreover, we had to rework the specification of a_NonceClient in this
operation to ensure that the specification is well defined and to make
proofs easier.