Ticket #1602: Rework of server_create_session_req_and_resp_sm
server_create_session_req_and_resp_sm had too many POs because it has a
lot of "if" in the implementation.
I reworked it by creating two locals operations, it permits to be more
readable and to reduce the number of POs.
I made some changes on other operations to make the proofs.