Ticket #1602: Complete preconditions of operations l_secure_channel_lost
We have to add some invariants in the preconditions because it's a local
operation.
There was a duplication of p_session : t_session_i in both operations.
Moreover, we iterate on every session, no matter if it's server or
client, so we have to remove the precondition about the nature of the
session.