Ticket #1602: Deleting set_session_orphaned
This operation did two things, removing the channel and maybe make a
relation in a_client_orphaned.
It was called only twice, in the operations l_secure_channel_lost.
After analyse, we can remove the call of set_session_orphaned in the
server local operation and just remove the channel from the session in
the implementation.
Moreover, in the client local operation, we know that
p_channel_config_idx : t_channel_config_idx, so we can add a relation in
a_client_orphaned without testing the nature of p_channel_config_idx.
To do that we can promote two operations from session_core_2 which do
exactly what we want, then we can delete set_session_orphaned which is
unused now.