Ticket #1602: Rework of client_secure_channel_lost_session_sm
The previous operation was not complete and was difficult to prove. We
needed to add some invariants in the loop and in the local operation.
This way of local operation is better for proving and is more readable.
In the implementation we can make the test of "l_dom = TRUE & l_channel
= p_lost_channel" before the call of the local operation to avoid some
useless calls.