Ticket #1602: Rework of server_secure_channel_lost_session_sm
As well as client_secure_channel_lost_session_sm, this operation was not
complete. We needed to add some invariants in the loop invariant and in
the local operation. Moreover, we needed to rework the local operation
to be more readable and to have easier expressions.
We can make the test of "l_dom = TRUE & l_channel = p_lost_channel"
before the call of the local operation to avoid useless calls.