Ticket #1602: Simplify specification of set_session_channel
This writing of a_channel means the same than the previous one, but is
better for proving, because it is closer to the machine invariant.
We don't need to specify how a_channel_nb_sessions is calculated, we
just need to give its final value which is the invariant.
This simplifies the specification, it's more readable and the proofs are
easier.

Note: This change necessitates to rename the channel parameter to avoid
a conflict with the quantified variable in the post-condition.