Ticket #1658: client wrapper / state machine: now actually disconnects the secure channel
Now that the low level API allows to disconnect the secure channel, the state machine now close the secure channel when the session is closed.

It also allows to manage closing only the secure channel when no session has been established (discovery services).
Client wrapper discovery services behavior:
- synchronous discovery service: automatically close the connection.
- asynchronous discovery service: API to close manually the connection.
Note: for the latter case this is the only way to do it as response callback thread cannot call it synchronously without a deadlock or possible incoherent state if disconnection is asynchronous (e.g. calling again discovery service on same connection).