Ticket #1666: CallContext: atomic counter shall be decremented for checking value
We were checking the value of counter using atomic get to determine if there was remaining copies.
But since it might be done concurrently it might lead not to deallocate the value with current code:
- get counter = 2 for both threads (e.g. T1 an T2) that then will decrement it: counter = 0 at the end without deallocation
The counter shall be decremented at same time the value is checked to guarantee the counter = 1 value is returned to 1 thread and deallocation occurs.